+++ lib/src/rhttp.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:rhttp/src/interceptor/interceptor.dart';
 import 'package:rhttp/src/model/cancel_token.dart';
 import 'package:rhttp/src/model/request.dart';
 import 'package:rhttp/src/model/response.dart';
@@ -18,6 +19,7 @@ class Rhttp {
   /// Use [send] if you already have a [BaseRhttpRequest] object.
   static Future<HttpResponse> request({
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     required HttpMethod method,
     required String url,
     Map<String, String>? query,
@@ -29,6 +31,7 @@ class Rhttp {
       requestInternalGeneric(RhttpRequest(
         client: null,
         settings: settings,
+        interceptor: parseInterceptorList(interceptors),
         method: method,
         url: url,
         query: query,
@@ -43,16 +46,19 @@ class Rhttp {
   static Future<HttpResponse> send(
     BaseRhttpRequest request, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
   }) =>
       requestInternalGeneric(RhttpRequest.fromBase(
         request: request,
         client: null,
         settings: settings,
+        interceptor: parseInterceptorList(interceptors),
       ));
 
   /// Makes an HTTP request and returns the response as text.
   static Future<HttpTextResponse> requestText({
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     required HttpMethod method,
     required String url,
     Map<String, String>? query,
@@ -62,6 +68,7 @@ class Rhttp {
   }) async {
     final response = await request(
       settings: settings,
+      interceptors: interceptors,
       method: method,
       url: url,
       query: query,
@@ -76,6 +83,7 @@ class Rhttp {
   /// Makes an HTTP request and returns the response as bytes.
   static Future<HttpBytesResponse> requestBytes({
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     required HttpMethod method,
     required String url,
     Map<String, String>? query,
@@ -85,6 +93,7 @@ class Rhttp {
   }) async {
     final response = await request(
       settings: settings,
+      interceptors: interceptors,
       method: method,
       url: url,
       query: query,
@@ -99,6 +108,7 @@ class Rhttp {
   /// Makes an HTTP request and returns the response as a stream.
   static Future<HttpStreamResponse> requestStream({
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     required HttpMethod method,
     required String url,
     Map<String, String>? query,
@@ -108,6 +118,7 @@ class Rhttp {
   }) async {
     final response = await request(
       settings: settings,
+      interceptors: interceptors,
       method: method,
       url: url,
       query: query,
@@ -123,12 +134,14 @@ class Rhttp {
   static Future<HttpTextResponse> get(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     CancelToken? cancelToken,
   }) =>
       requestText(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.get,
         url: url,
         query: query,
@@ -140,12 +153,14 @@ class Rhttp {
   static Future<HttpTextResponse> getText(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     CancelToken? cancelToken,
   }) =>
       requestText(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.get,
         url: url,
         query: query,
@@ -157,12 +172,14 @@ class Rhttp {
   static Future<HttpBytesResponse> getBytes(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     CancelToken? cancelToken,
   }) =>
       requestBytes(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.get,
         url: url,
         query: query,
@@ -174,12 +191,14 @@ class Rhttp {
   static Future<HttpStreamResponse> getStream(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     CancelToken? cancelToken,
   }) =>
       requestStream(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.get,
         url: url,
         query: query,
@@ -192,6 +211,7 @@ class Rhttp {
   static Future<HttpTextResponse> post(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     HttpBody? body,
@@ -199,6 +219,7 @@ class Rhttp {
   }) =>
       requestText(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.post,
         url: url,
         query: query,
@@ -212,6 +233,7 @@ class Rhttp {
   static Future<HttpTextResponse> put(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     HttpBody? body,
@@ -219,6 +241,7 @@ class Rhttp {
   }) =>
       requestText(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.put,
         url: url,
         query: query,
@@ -232,6 +255,7 @@ class Rhttp {
   static Future<HttpTextResponse> delete(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     HttpBody? body,
@@ -239,6 +263,7 @@ class Rhttp {
   }) =>
       requestText(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.delete,
         url: url,
         query: query,
@@ -252,12 +277,14 @@ class Rhttp {
   static Future<HttpTextResponse> head(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     CancelToken? cancelToken,
   }) =>
       requestText(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.head,
         url: url,
         query: query,
@@ -270,6 +297,7 @@ class Rhttp {
   static Future<HttpTextResponse> patch(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     HttpBody? body,
@@ -277,6 +305,7 @@ class Rhttp {
   }) =>
       requestText(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.patch,
         url: url,
         query: query,
@@ -290,6 +319,7 @@ class Rhttp {
   static Future<HttpTextResponse> options(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     HttpBody? body,
@@ -297,6 +327,7 @@ class Rhttp {
   }) =>
       requestText(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.options,
         url: url,
         query: query,
@@ -310,6 +341,7 @@ class Rhttp {
   static Future<HttpTextResponse> trace(
     String url, {
     ClientSettings? settings,
+    List<Interceptor>? interceptors,
     Map<String, String>? query,
     HttpHeaders? headers,
     HttpBody? body,
@@ -317,6 +349,7 @@ class Rhttp {
   }) =>
       requestText(
         settings: settings,
+        interceptors: interceptors,
         method: HttpMethod.trace,
         url: url,
         query: query,